convert Val* fingers to int regs
This is mostly a no-op bookkeeping change, converting our cleanup
fingers from pointing into regs[] to register IDs themselves, following
the usual convention of NA meaning none.

But this does allow us a future flexibility, letting us think of regs[]
as the inter-instruction register state, tracking intra-instruction
dst,x,y,z registers a little more carefully.

Under the old system, if you alias id and x into register 7, which Val
do we write in regs[7]?  Neither is exactly correct, and writing either
one breaks calls to r(the other).  But long term, the right answer will
be id... that's the Val that's in the register as the next instruction
begins.

So rx can answer "which register holds Val x?" during the instruction
even when we alias x and the destination, when regs[rx] == id, not x.

Spinning this off a precursor CL, but probably no point landing it
unless the CL that builds on this works out.

Change-Id: I528c46ec2d00a965125749562ef10b76f973a3cd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/284955
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Mike Klein <mtklein@google.com>