stack-aware codegen?
It occurs to me that now that we have some values living on the stack,
we might actually want to change our codegen based on what's in
registers or what's not.

E.g.

    a->vaddps(dst(), r(x), any(y))

is great if x is in a register, but it'd have been better as

    a->vaddps(dst(), r(y), any(x))

if x is on the stack.  I've made every symmetric op make this choice.

Change-Id: I1c65940d8e1a53d9dba44dd3a5f3e443517c8e20
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/284983
Commit-Queue: Mike Klein <mtklein@google.com>
Reviewed-by: Herb Derby <herb@google.com>