stack-aware codegen?
It occurs to me that now that we have some values living on the stack,
we might actually want to change our codegen based on what's in
registers or what's not.

E.g.

    a->vaddps(dst(), r(x), any(y))

is great if x is in a register, but it'd have been better as

    a->vaddps(dst(), r(y), any(x))

if x is on the stack.  I've done add_f32 and mul_f32 as a demo here, but
the same logic is true for all symmetric ops, which I think means also
add_i32, mul_i32, bit_and, bit_or, bit_xor, eq_i32, eq_f32, and neq_f32.

I think it'll be good to get _this_ thought through before we think
about aliasing dst() to argument registers.

Change-Id: I1c65940d8e1a53d9dba44dd3a5f3e443517c8e20