general dst() aliasing
Clarity refactors: rename alias_reg() to try_alias() and add in_reg().

Let dst() hint what register to try_alias(), and use this anywhere
aliasing is safe and we have an argument that's either already in a
register or is guaranteed to be pulled back into one, e.g r(x).

Cases like vsqrtps(dst(), any(x)) are somewhat interesting, in that if x
is in a register, it is a good thing to alias, but if it's on the stack,
it's better to just vsqrtps directly into the new register.

Change-Id: I05ecb78f660e9e4ea0df44f03e5de03b85a49cc7