Update patch set 4
Patch Set 4:

(As discussed online)

The eventual goal of the investigation would be to adopt some absl code within Skia itself. We won't be able to do that until we actually resolve the Android issues. Of course, we could still disable Android in the short-term and submit this CL, then separately we could focus on solving the Android issue. Let's see how complicated it will be to fix Android and then we can decide.

Patch-set: 4
Reviewer: Gerrit User 11059 <11059@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0