rework constants
Splat instructions no longer do anything at runtime, merely
making sure at compile time that the constant will be available.

Then we can locate(Val, Reg*, Label**) to find a given value, and use
what it returns to either create an x86 Operand for that value or load
it into a Reg as needed.  Since constants are already in-memory, they're
also no-ops to spill.

This lets us simplify and delete lots of code, and we now only call
optimize() once for all backends, so it's a big deal for overhead.

Ideas all drafted, but not quite working.

Commit: false
Change-Id: I78a54feaf8b18be26efe73dffda903a7a9e31daa