Update patch set 11
Patch Set 11:

Looks like we're tracking all the constants correctly, and we know what instructions reference them, and what offsets they should be at, but when we update the offsets for some of the comparison instructions we end up one byte off:

    0x10729c011: vmovups (%rsi), %ymm0

    0x10729c015: vmovups 0x2c7(%rip), %ymm2           ; 5.0f
    0x10729c01d: vcmpleps %ymm0, %ymm2, %ymm1
    0x10729c022: vpand  0x21a(%rip), %ymm1, %ymm1     ; 20

    0x10729c02a: vmovups 0x232(%rip), %ymm4           ; 4.0f
    0x10729c032: vcmpltps %ymm0, %ymm4, %ymm3
    0x10729c037: vpand  0x285(%rip), %ymm3, %ymm3     ; 10

    0x10729c03f: vcmpleps 0x15d(%rip), %ymm0, %ymm5   ; 3.0f
    0x10729c048: vpand  0x254(%rip), %ymm5, %ymm5     ; 8

    0x10729c050: vcmpltps 0x22c(%rip), %ymm0, %ymm6    ; 2.0f
    0x10729c059: vpand  0x183(%rip), %ymm6, %ymm6      ; 4

    0x10729c061: vcmpneqps 0x15b(%rip), %ymm0, %ymm7   ; 1.0f
    0x10729c06a: vpand  0x192(%rip), %ymm7, %ymm7      ; 2

    0x10729c072: vcmpeqps 0x10a(%rip), %ymm0, %ymm0    ; 0.0f
    0x10729c07b: vpand  0x1a1(%rip), %ymm0, %ymm0      ; 1


In this snippet the addresses for 5.0f (0x10729c01d + 0x2c7 = 0x10729c2e4) and 4.0f (0x10729c032 +  0x232 = 0x10729c264) are correct, but the others are not, e.g. 3.0f (0x10729c048 + 0x15d = 0x10729c1a5, but we want 0x10729c1a4).

Notice the comparisons with registers are good and the comparisons with memory directly have this off-by-one problem.  So I think maybe there's nothing wrong with this CL per se, but it's exposed a bug in the assembler in how cmpps and memory arguments interact.

Patch-set: 11
Reviewer: Gerrit User 5045 <5045@5086850b-e599-37d2-8ec8-ff16aeccbf02>