Update patch set 2
Patch Set 2:

This is the bug... these marked instructions are encoded as

   prefix - opcode - mod/rm - offset - imm-byte

and when we go back to fix up the offset, we're doing it
relative to the end of the offset, not relative to the end
of the whole instruction.  We end up one byte off, and that's
not good. 

A quick fix is to disallow Operand on them, but it might be
worth seeing if we can just fix it first somehow.

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5045 <5045@5086850b-e599-37d2-8ec8-ff16aeccbf02>