Update vulkan backend texture creation to use ganesh abstractions of vk resources.
This allows us to remove all the manual creation of vulkan objects in
the backend texture code. It also uses our command buffer tracking system
to know when to free and delete the vulkan resources.

Ideally we would be able to share even more code with uploading to
textures and command buffer submission, but more suibtle refactoring
would be needed to keep everything running smooth and not overly complex.
This changes at least allows us to move forward with current plans on
updating the backend texture data upload APIs.

Change-Id: I87aba838615b0c4538513fdc753fd8892c30948e