Optimize GrTessellatePathOp's code to emit inner triangles
Previously we used a naive algorithm to generate inner polygon
triangulations with a "middle-out" topology. We copied all endpoints
to a new array, then recursed on this array.

This CL adds a "GrMiddleOutPolygonTriangulator" class that
accomplishes the same thing in 1/5th the time using a small O(log N)
stack.

Change-Id: I3a7059e5d133a730b7084a17d8fbaaa3aaa81336