Fix bug in aa stroke clip path
When trying to chop a quadratic along the right edge of a clip,
the computed t value was so close to 1.0 that our chopper method
returns false (we would have needed doubles to detect this).

To handle this, pin the X values to the right edge, so that we at least
maintain the contract that we are clipped.

Bug: 1070835
Change-Id: Ifdc59f97c7f5c32b321647f6739b37b33ce801c9