Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> I'm also ok if we just tables strings for now. I like 'em, but user-specified ints work too (at least for a while, while we flesh-out other aspects of this). Not asking to go back to ints, just saying we can probably live with 'em in a pinch.

Overall, I like the readability that strings in sksl produce, and I like the relative simplicity of this CL in terms of letting the lower levels deal with just numbers.

I would not really want to go back to using 'ints', since it seems likely that additional aspects of runtime sksl will benefit from whatever user-provided naming system we come up with. 

Perhaps it would be easier to have a stable hash function used for this purpose that we can rely on for serialization? Unless it turns out not to be too much extra work for SkVertices to remember the names associated with its hashes, that it uses during serialization.

Patch-set: 7
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>