Retry drawing the simplified shape
Previously, functions like GrRTC::drawRect or drawRRect would check
for common cases and then fallback to creating a GrStyledShape and
calling drawShapeWithPathRenderer. This would always use path rendering,
even if the analysis that GrStyledShape did was able to change the style.

This adds one additional retry to go through drawShape() before the
settling on path rendering.

Change-Id: I1dcf218f9bf83d40dad992352bcd4a6e74b01fad