Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> PS7 stores the names in the vertices object, and serializes those (then reconstructs the hashes when deserializing). Hash function no longer needs to be stable or consistent. (The upcoming SkSL change will automatically have similar behavior, as we serialize the SkSL itself). At this point, hash collisions are a risk, but:
> 
> 1) Accidental collisions are going to be incredibly unlikely. A typical user of the API will have 1 custom matrix (local-to-<wherever they do lighting>). Ambitious clients may have a handful. Depending on backend, shader limits may even put a hard upper limit on the number of uniform matrices we can upload (<100).
> 
> 2) Intentional collisions won't be difficult to cause, but no collision (intentional or otherwise) is actually dangerous - you'll just get the wrong matrix, and draw incorrectly.

Yeah, I think this sort of setup is ok as long as we merely have to change some code should we notice it's become a problem.  I'm not sure how we'd do that if we've serialized by those hashes except by declaring that serialized data unreadable.

It can be worth playing around with the numbers to make sure you're comfortable with P(collision), e.g. http://davidjohnstone.net/pages/hash-collision-probability.

Patch-set: 7