Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> We do have some animated GIF samples (e.g. the nyan cat here: https://skia.org/user/modules/skottie).
> 
> Is the plan for SkottieKit to always be a minimal build, of would it make sense to support multiple configs? E.g. SkottieKit/Full, SkottieKit/Min.
> 
> I'm fine to move in the Min direction by default, but there seems to be some value in having a full-featured build.  Or is the idea to just use CK in that case?

SkottieKit at the moment is the experiment to get the absolute smallest code size possible that still functions. If in the future, we continue to need a dedicated Skottie build, then I'm fine with adding options and more features, with the default being as stripped down as possible.

Patch-set: 2
Reviewer: Gerrit User 5395 <5395@5086850b-e599-37d2-8ec8-ff16aeccbf02>