cleanup device bounds calculation
We can store the so that they are relative to (0, 0), instead of the
initial origin. This simplifies much of the math for calculating the
device bounds. And, it remove the need for the inverse matrix in
GrTextBlob.

As a note, it looks like GrAtlasTextOp::init() may be calculating the
wrong bounds because it doesn't seem to respect needsTransform()
correctly. I'll check this out in the next CL.

Change-Id: Ia053b8b11ef8bbae5cc9e7bc547b918401519027
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/285839
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>