Refactor geometry union capabilities out of GrStyledShape
The geometry union part of GrStyledShape is now held in GrShape. For the
most part, GrShape is entirely style agnostic and focuses on storing
the various types of geometry, and destructing them gracefully. It also
provides a public API that unifies functionality across all shape types,
such as contains() and bounds().

GrStyledShape now just owns a GrShape and a GrStyle, and handles the
additional simplification logic that relies on knowing the effects of
the style on the draw. This is where GrShape makes some allowances for
style. Its simplify() function accepts flags that enable/disable various
simplification optimizations. Currently these are designed around
what is needed to respect path effects and stroking behaviors in
GrStyledShape. The main other user of GrShape (the new clip stack) will
always provide all flags since it treats every shape as if it were
simply filled.

Several other related refactorings were taken at the same time:
1. The implementations for asNestedRects, asRRect, etc. were moved out
   of the header and into the cpp file for GrStyledShape.
2. GrRenderTargetContext relies on GrStyledShape for its stroke rect
   fallbacks.
3. GrShape can hold points, lines, and rects explicitly. This let me
   simplify the stroke reasoning.

Change-Id: I9fe75613fee51c30b4049b2b5a422daf80a1a86e