Add a faster, cleaner raw path iterator
Adds a zero-copy range-for path iterator and rewrites the guts of
SkPath::RawIter to use the new iterator. This iterator temporarily
exists as SkPath::RawIterImpl, but will soon be moved to
SkPathPriv::Iter once we update all of RawIter's call sites.

Using this instead of the copy-heavy SkPath::Iter nearly doubles
performance on GrTessellatePathOp::prepareOuterCubics. The resulting
code is also more neat and keeps the iteration variables (verb and
pts) scoped strictly inside the for-loop.

Change-Id: I64b929d7015a349dd2c64744dc48132a286778b0