Update patch set 8
Patch Set 8: Code-Review+1

(5 comments)

Would it make sense to also have a row/column or reference GM that should produce the same image, but works by rendering the src with the tile mode over the test area, extracting the dst rect from that, blurring, and then drawing?

That would only be useful if you find it harder to reason about what the expected behavior of the blurs should be, but if that's not being a problem, probably isn't worth the effort.

Patch-set: 8
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1