Add a faster, cleaner path iterator
Adds a zero-copy "RangeIter" class and rewrites the guts of
SkPath::RawIter to use RangeIter instead. The new iterator can be
used with a standard range-for loop:

 for (auto [verb, pts, weights] : skPath) {
     ...
 }

Using this instead of the copy-heavy SkPath::Iter nearly doubles
performance on GrTessellatePathOp::prepareOuterCubics. The resulting
code is also more neat and keeps the iteration variables
(verb, pts, weights) scoped strictly inside the for-loop.

Change-Id: I64b929d7015a349dd2c64744dc48132a286778b0