Add a faster, cleaner path iterator
Adds a zero-copy "RangeIter" class and rewrites the guts of
SkPath::RawIter to use RangeIter instead. Adds functionality to
SkPathPriv to enable iteration with a standard range-for loop:

 for (auto [verb, pts, weights] : SkPathPriv::Iterate(skPath)) {
     ...
 }

Once the usage cases of RawIter are updated, this iterator will be moved
to SkPathPriv.

Using this instead of the copy-heavy SkPath::Iter nearly doubles
performance on GrTessellatePathOp::prepareOuterCubics. The resulting
code is also more neat and keeps the iteration variables
(verb, pts, weights) scoped strictly inside the for-loop.

Change-Id: I64b929d7015a349dd2c64744dc48132a286778b0