Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> > Patch Set 12: Commit-Queue+1
> > 
> > > Patch Set 8:
> > > 
> > > seems like the client can reference fData directly, and therefore us it (incorrectly) with uninitialized values (if they haven't set all of their uniforms). Do we care?
> > 
> > I'm not real bothered by that - I think that clients may have different approaches to gathering and setting all of the data, so I was leaning towards making it less opinionated.
> > 
> > We could add utility functionality that lets them at least validate that they set all inputs (track which uniforms have been assigned, then have a validate() that checks?).
> 
> Hmmm, does the client really know how fData is packed anymore? Why not variants like...
> 
> builder b;
> b["lightPos"] = ...;
> 
> <-- b.makeShader(children);
> or
> <-- effect->makeShader(b, children);

If we go down this road, it's a thin line between the effect and the builder (we could just have the effect itself allocate the "working" blob, and provide the same interface for setting values - then the makeShader API doesn't need the inputs at all). As it stands now, the builder is entirely optional - the client can do everything it's doing with public API. (Though adding some extra syntactic sugar to the builder or the effect to remove the indirection would maintain that). If I were to do that, I think I'd go with the first verion (makeShader on the builder) and also have it own a list of the children that can be set via name, for consistency. I can mock that up...

Patch-set: 13
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0