Update patch set 9
Patch Set 9:

(3 comments)

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > I feel like I'm getting close to what we want in PS 7.  Seems like my only issue remaining is that the overall bot task is green and not red?
> 
> I read in the docs,
> 
> 	// The Task Driver as a whole fails when the root-level step fails, or
> 	// when EndRun() catches a panic.
> 
> but that doesn't seem to happen?  Should I be explicitly td.Fatal()ing at the end if there were any failures?

Yes, you have to explicitly fail the task driver.  There are a number of instances of non-fatal errors, eg. checking for existence of a file or directory on an Android device before spending the time to push.  That step might return an error if the directory doesn't exist, but it's up to the caller to determine whether to bubble the error up or handle it.

Patch-set: 9
Reviewer: Gerrit User 5015 <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>