Dawn: Use shared copy encoder to create backend texture
In onCreateBackendTexture(), a new CommandEncoder was being created to
copy from the staging buffer to the result texture. Following the
staging manager rewrite, this copy encoder was being submitted before
the staging buffer was flushed, causing texture uploads to fail.

Use fCopyEncoder to ensure the staging buffer is unmapped before the
commands are run.

Bug: chromium:1071073
Change-Id: I3e8a77f3356671411f01dbb8017a7997e7c9349b