Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Looks like there's a case where fInverted doesn't get initialized?

I think it was just because I never assigned it during setPath(), since inverted() uses the path's fill type directly if the shape is a path. But operator= just copied all the fields over, so it'd get the uninitialized garbage. We'll see if the bots stop complaining

Patch-set: 4
Label: Commit-Queue=0