Update patch set 3
Patch Set 3:

(2 comments)

Do you plan on having another class that sits on top of this that manages allocations for multiple descriptors. Like say I have a pipeline that needs 3 SRVs, we need something that is able to reserve a contiguous block of 3 descriptors in a heap for them and later on know to make those blocks available again later.

Patch-set: 3
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>