Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (2 comments)
> 
> Do you plan on having another class that sits on top of this that manages allocations for multiple descriptors. Like say I have a pipeline that needs 3 SRVs, we need something that is able to reserve a contiguous block of 3 descriptors in a heap for them and later on know to make those blocks available again later.

I assumed there'd be another class -- either the ResourceProvider, or the Pipeline maybe? -- that uses this to "allocate" a descriptor, and then set it to be a SRV or CBV using the methods in ID3D12Device (e.g. CreateShaderResourceView). So that class could manage contiguous blocks as well. Or we can add a new method that returns a set of descriptors.

I also sort of see us using this in two different ways. The first is an allocation for the lifetime of the GrContext, for things like RTVs and DSVs. That's where we'd use both allocate() and free() as we create and destroy surfaces, much like Vulkan creates and destroys VkImageViews. The second is more like how GrVkDescriptorPool is used in Vulkan -- we use it to allocate shader descriptors and then just reset it (in the pool sense, not the sk_sp sense) once the command list is finished. So you wouldn't need to call free() three times.

You can look at my d3d clear change (*** in progress, not ready for review ***) for an example of how I use it: https://skia-review.googlesource.com/c/skia/+/286042.

Patch-set: 3
Reviewer: Gerrit User 5065 <5065@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0