Adding some intrinsics to the interpreter
Includes all of the vectorized comparisons, as well as the selecting
and lerping variants of mix. The previous implementations of lerping mix
were available based on the compiler front-end, this moves them to the
byte code generator, so they're available in runtime effects (not just
generic interpreter programs).

Change-Id: I88d10c8a18a7e63e49acac8181b687e2b40ad66b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/287008
Reviewed-by: Mike Klein <mtklein@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>