Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Feels like the file system is leaking through the SkData abstraction.  Let’s rewrite this to not use SkData?

I'm still not sure what you mean, which SkData abstraction? The SkData abstraction is that you can read in some number of unsigned char through a pointer. The backing of the SkData can hold onto whatever resources it needs to. It seems a bit silly to force read all the bytes of a file up front into memory on Windows for behavior that is absolutely normal (and expected) on Windows. The doc for SkData::MakeFromFileName doesn't say anything about allowing external modification of the file. Perhaps just because I've worked with Windows and this particular issue enough (I recognized more or less what was happening from the initial error message) I don't find this any more 'spooky' than having any type which may wrap another type which may hold resources. The SkData holds a pointer to a file HANDLE which is locking the file.

And of course SkData doesn't mmap on Windows, there is no mmap on Windows. I think I know what you meant, but just having a FILE* open is the same thing, the file will be locked. It is the source of much frustration on Windows.


In any event, I noticed that my change precludes writing the file out when it doesn't already exist, which it seems this did before. Fixing that sidesteps the Windows locking things issues (by putting the reading and writing in separate scopes anyway).

Patch-set: 2
Label: Commit-Queue=0