Qualify constant strings with constexpr as appropriate.
When at global scope, also removed the `static` qualifier, as
`constexpr` will already prevent external linkage. (At local scope,
`static` was left as-is. This prevents code-size growth in debug builds,
although in release it appears to generate identical code with or
without `static`.)

The existing call sites such as `const char* kText = "foo";` did not
actually prevent kText from being changed to a new value. Using
constexpr is safer and allows the compiler to know with certainty that
kText cannot be changed. For instance, after constexpr was added, the
compiler was able to identify an incorrect format string in
skpbench.cpp.

Some inspiration for these changes was taken from http://go/totw/140.

Change-Id: Ia30a7879a62c4af487f1eff62cc2eddc54addbac