Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Think it makes sense to back this with std::vector<bool> now?  Or just use std::vector<bool> where we use SkBitSet?

Well, there are at least two differences. The first being that SkBitSet is a fixed size (doesn't need to grow or shrink after construction) which is a nice property not to need to reason about. The second is that now that we have a findFirst I don't know how to do that in any way efficiently with std::vector<bool> since it doesn't provide it directly and you can only test a bit at a time. Maybe the compiler could try to see through all that and 'auto-vectorize' it but I would want to be reasonably confident that happens. This findFirst thing is actually what most of the code changes are about here (found because of the added testing), other than fixing that this is actually a rather simple change (I just wanted to rename things to be like std::bitset, since Herb mentioned it last time I touched this file and I said I'd get back to it).

Patch-set: 5
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>