Fix path mask caching regression
Adding the explicit rect to the path mask key in:

https://skia-review.googlesource.com/c/skia/+/286902 (Add path bounds to SW path mask key)

prevented the path mask from being reused more generally.

This CL just adds the width & height to the key which allows reuse but still prevents proxy size mismatches due to numerical inaccuracies.

For the path_mask_cache GM we have:

                 Num Generated   Num Cache Hits

Current ToT:          24               0
prior to bad CL:      18               6
with this CL:         18               6

Change-Id: I6f5c847e6f14c80013d4f315ef8a9a16c737295b