Update patch set 5
Patch Set 5:

Partially it's because this was exactly the implementation I was looking for, and I don't necessarily know what I'll get with vector<bool>. Partially it's because I find vector<bool> to be esthetically weird for a bitset, but more practically it doesn't provide find_first and iterating over each bit seems like an inefficient solution (though, sure, let's test it). I'd be more in favor of using boost::dynamic_bitset, though that's another can of worms.

Patch-set: 5