Update patch set 2
Patch Set 2:

I'm not sure this is the right answer. From a UX perspective, it'd be better not to have the option available at all if we can't do it. 

How about changing line 1700 to 
if (ctx && ctx->maxSurfaceSampleCountForColorType(kRGBA_8888_SkColorType) > 1) {
?

Patch-set: 2
Reviewer: Gerrit User 5065 <5065@5086850b-e599-37d2-8ec8-ff16aeccbf02>