Update patch set 1
Patch Set 1:

Making this change so that the follow on change to separate create and update can remove some complexity in code and the contract. If this was a perf gain at all, it was probably a very very small one.

Patch-set: 1
Reviewer: Gerrit User 5230 <5230@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Work-in-progress: false