Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > complexclip4 is the one that relies on Android's device restriction. We could wait to remove it until hearing a decision, land it now anyways, or split out the complexclip4 deletion until later.
> 
> Don't we need to wait on a decision before any of this?

Probably, I wanted this change to make it easier to test correctness for the new clip stack, but I don't have to land this in order to do that.

I could update the CL to instead guard the extra ops with the SK_SUPPORT_DEPRECATED_CLIPOPS define, and then we set up a bot or two that doesn't add that flag. Then the new clip stack could be active and tested on those bots, while we use SkClipStackDevice everywhere that still defines the deprecated clipops flag.

Patch-set: 1
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0