Update patch set 4
Patch Set 4:

Are we not going to do this for all the other backends?

On a side note, when adding the D3D cache I started thinking all our caches for different backends are basically the same. The only real difference is how to make the GrProgramDesc which is already a virtual on caps. And then the actual making of the program/pipeline. This can easily be a gpu virtual since I think they all take the same inputs.

If we just had one shared cache system a change like this wouldn't need to replicate to all the backends.

Patch-set: 4
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>