Update patch set 1
Patch Set 1:

So looks like this is still not able to get the vulkan library on the device. I see the change you man in android.py, but when I look at the execution_details for nanobench on the bot I don't see anything about setting the LD_PRELOAD variable. Does that occur in a different step?

Patch-set: 1
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>