Update patch set 8
Patch Set 8:

This is a good driver to force us to fix serialization. In particular we need something at the beginning of the font data to state what format it is and how we would prefer it to be deserialized (by FreeType, or the default, or GDI, or DirectWrite, or CoreText, or whatever this builder is). We've needed that for the SkTestFont and SkTestSVGFont for ages but have been getting by without doing anything by only using them as defaults and not using anything else at the same time.

Patch-set: 8
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>