Improve thin horiz/vertical dashed lines by not forcing width to 1.
It looks like a bug slipped in quite a while back here:
https://codereview.chromium.org/1092793006/diff/30002/src/gpu/effects/GrDashingEffect.cpp

where a width snapping that was supposed to apply to non-AA lines was
changed to apply to AA lines.

Adds GM that tests < 1 pixel wide dashed horiz/vertical lines.

The lines look better with the change but are still a little too thick,
depending on the subpixel offset.

BUG: chromium:1049028

Change-Id: I45734f5ef55548b62c188d9f55d3150c00059eed
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/288628
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>