Dawn readpixels fixes.
The dawn backend was calling GrDawnGpu::flush() in a few places (in
particular, for readpixels).Since GrGpu::submitToGpu() is now
responsible for unmapping the staging buffers prior to command
submission, internally calling GrDawnGpu::flush() is error-prone,
since it requires you to unmapStagingBuffers() manually.

The fix is to get rid of GrDawnGpu::flush(), and put its guts into
GrDawnGpu::onSubmitToGpu(). Then we call submitToGpu(true) where we
used to call flush().

Also, make the backend responsible for moving staging buffers from the
active to the busy list. We do it just before calling mapAsync(),
since the callback may return right away and we want the buffer to be
in the busy state at that point.

Change-Id: I99cfa5e7ea3a454b0383a30483fa93795191347a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/288516
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Stephen White <senorblanco@chromium.org>