Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> I like the inner class pattern better too; the matrix convolution processor has to expose all of its members right now the same way the old gaussian processor did.

Once it's converted to inner class I really want to delete the GenKey() function on Impl and just have the outer class implement the key. I think the original reason for GenKey() was that the GrGLSLFragmentProcessor class was responsible for the code gen that the key represents. But once they're best buddies as inner/outer it seems really unnecessary (if it ever was).

Patch-set: 3
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>