Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> My only thought here would be to maybe land the GM first so we can easily see what the fix actually changes. I also wonder if it is worth doing a non-aa version of the GM to see if this fix changes what those draw at all (though I get we may get annoying snapping diffs on different gpus).
> 
> Regardless I'm also happy to just land this as is.

The AA version looks bad without the fix, verified that locally. Originally I made the GM draw AA and non-AA. The reason I removed non-AA is that GPU and CPU behave differently in a way that would make this hard to triage. Because the GPU always makes the lines 1 pixel thick we draw all the rows/columns. The CPU only draws for some subpixel offsets, presumably when the pixel centers are covered. So they look very different. I think the CPU is more correct, consistent with other geometries. As is it stands now though, the CPU looks "worse" at a surface level than GPU because lines drop out. It's probably something worth making more consistent across backends, but we'd probably want a more complicated GM that clearly identifies which lines we expect to draw something and which should not. Maybe something like this: At each subpixel offset determine whether it should or should not draw and set the paint color to blue if it should and red if not. First draw at pixel center in the opposite color and then at the subpixel offset in the chosen color. If it is correct you should have all blue lines and if its wrong you see red lines. Anyway, seems like a different GM for future work.

Patch-set: 2
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0